third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister-controller.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var worker_url = 'resources/simple-intercept-worker.js';

async_test(function(t) {
    var scope = 'resources/unregister-controller-page.html?load-before-unregister';

    service_worker_unregister_and_register(t, worker_url, scope).then(t.step_func(function(worker) {
        worker.addEventListener('statechange', t.step_func(on_state_change));
    }));

    function on_state_change(event) {
        if (event.target.state != 'active')
            return;
        with_iframe(scope, t.step_func(function(frame) {
            var w = frame.contentWindow;
            assert_true(w.navigator.serviceWorker.controller instanceof w.ServiceWorker,
                        'document should load with a controller');
            w.fetch_url('simple.txt').then(t.step_func(function(response) {
                assert_equals(response, 'intercepted by service worker', 'controller should intercept requests');
                t.done();
            }), unreached_rejection(t, 'fetch failed'));
        }));
    }
}, 'Unregister does not affect existing controller');

async_test(function(t) {
    var scope = 'resources/unregister-controller-page.html?load-after-unregister';

    service_worker_unregister_and_register(t, worker_url, scope).then(t.step_func(function(worker) {
        worker.addEventListener('statechange', t.step_func(on_state_change));
    }));

    function on_state_change(event) {
        if (event.target.state != 'active')
            return;
        navigator.serviceWorker.unregister(scope).then(t.step_func(function() {
            with_iframe(scope, t.step_func(function(frame) {
                var w = frame.contentWindow;
                assert_equals(w.navigator.serviceWorker.controller, null, 'document should not have a controller');
                w.fetch_url('simple.txt').then(t.step_func(function(response) {
                    assert_equals(response, 'a simple text file\n', 'requests should not be intercepted');
                    t.done();
                }), unreached_rejection(t, 'fetch failed'));
            }));
        }));
    }
}, 'Unregister prevents control of subsequent navigations');
</script>
